+++ src/node/serverPluginModuleRewrite.ts
@@ -43,6 +43,9 @@ export const moduleRewritePlugin: Plugin = ({ app, watcher, resolver }) => {
     `window.process = { env: { NODE_ENV: 'development' }}\n` +
     `</script>\n`
 
+  const scriptRE = /(<script\b[^>]*>)([\s\S]*?)<\/script>/gm
+  const srcRE = /\bsrc=(?:"([^"]+)"|'([^']+)'|([^'"\s]+)\b)/
+
   app.use(async (ctx, next) => {
     await next()
 
@@ -58,20 +61,29 @@ export const moduleRewritePlugin: Plugin = ({ app, watcher, resolver }) => {
       } else if (ctx.body) {
         await initLexer
         let hasInjectedDevFlag = false
-        ctx.body = html!.replace(
-          /(<script\b[^>]*>)([\s\S]*?)<\/script>/gm,
-          (_, openTag, script) => {
-            // also inject __DEV__ flag
-            const devFlag = hasInjectedDevFlag ? `` : devInjectionCode
-            hasInjectedDevFlag = true
+        const importer = '/index.html'
+        ctx.body = html!.replace(scriptRE, (matched, openTag, script) => {
+          const devFlag = hasInjectedDevFlag ? `` : devInjectionCode
+          hasInjectedDevFlag = true
+          if (script) {
             return `${devFlag}${openTag}${rewriteImports(
               script,
-              '/index.html',
+              importer,
               resolver
             )}</script>`
+          } else {
+            const srcAttr = openTag.match(srcRE)
+            if (srcAttr) {
+              // register script as a import dep for hmr
+              const importee = cleanUrl(slash(path.resolve('/', srcAttr[1])))
+              debugHmr(`        ${importer} imports ${importee}`)
+              ensureMapEntry(importerMap, importee).add(importer)
+            }
+            return `${devFlag}${matched}`
           }
-        )
+        })
         rewriteCache.set(html, ctx.body)
+        return
       }
     }
 
@@ -123,7 +135,8 @@ function rewriteImports(
       const currentImportees = new Set<string>()
       importeeMap.set(importer, currentImportees)
 
-      imports.forEach(({ s: start, e: end, d: dynamicIndex }) => {
+      for (let i = 0; i < imports.length; i++) {
+        const { s: start, e: end, d: dynamicIndex } = imports[i]
         let id = source.substring(start, end)
         let hasLiteralDynamicId = false
         if (dynamicIndex >= 0) {
@@ -136,7 +149,7 @@ function rewriteImports(
         if (dynamicIndex === -1 || hasLiteralDynamicId) {
           // do not rewrite external imports
           if (isExternalUrl(id)) {
-            return
+            break
           }
 
           let resolved
@@ -159,7 +172,8 @@ function rewriteImports(
             // append .js for extension-less imports
             // for now we don't attemp to resolve other extensions
             if (!/\.\w+/.test(pathname)) {
-              pathname += '.js'
+              const file = resolver.requestToFile(pathname)
+              pathname += path.extname(file)
             }
             // force re-fetch all imports by appending timestamp
             // if this is a hmr refresh request
@@ -189,7 +203,7 @@ function rewriteImports(
         } else {
           console.log(`[vite] ignored dynamic import(${id})`)
         }
-      })
+      }
 
       // since the importees may have changed due to edits,
       // check if we need to remove this importer from certain importees